deps: updates wazero to 1.0.0-pre.3 (#7)
This updates [wazero](https://wazero.io/) to [1.0.0-pre.3](https://github.com/tetratelabs/wazero/releases/tag/v1.0.0-pre.3). This is the last release that will build with Go 1.17.

Notably, this improves performance and changes host function syntax.

+++ go/wasmrs/internal/frames/request_payload_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestRequestPayloadFragment(t *testing.T) {
-	const maxFrameSize = 1024 + FrameHeaderLen + 3
+	const maxFrameSize = uint32(1024 + FrameHeaderLen + 3)
 	data := make([]byte, 1024*1024)
 	metadata := make([]byte, 1024*1024)
